+++ src/vec3.h
@@ -3,7 +3,7 @@
 #include <cmath>
 
 struct Vec3 {
-  float x, y, z;
+  double x, y, z;
 
   constexpr Vec3 operator+(const Vec3& other) const {
     return Vec3{x + other.x, y + other.y, z + other.z};
@@ -13,23 +13,23 @@ struct Vec3 {
     return Vec3{x - other.x, y - other.y, z - other.z};
   }
 
-  constexpr Vec3 operator*(float k) const {
+  constexpr Vec3 operator*(double k) const {
     return Vec3{x * k, y * k, z * k};
   }
 
-  constexpr Vec3 operator/(float k) const {
+  constexpr Vec3 operator/(double k) const {
     return Vec3{x / k, y / k, z / k};
   }
 
-  constexpr float norm() const {
+  constexpr double norm() const {
     return std::sqrt(x*x+y*y+z*z);
   }
 
-  constexpr float norm2() const {
+  constexpr double norm2() const {
     return x*x+y*y+z*z;
   }
 
-  friend constexpr Vec3 operator*(float a, const Vec3& b) {
+  friend constexpr Vec3 operator*(double a, const Vec3& b) {
     return b * a;
   }
 };
